+++ src/core/services/core-generate-request-body.service.ts
@@ -4,7 +4,10 @@ import { ContextGetDefaultService } from '../../context/context.module';
 import { Payload } from '../../payload/payload.module';
 
 export const CoreGenerateRequestBody = {
-  call: ({ context, ...payloadOptions }: any, configuration: Configuration) => {
+  call: (
+    { context, ...payloadOptions }: { [key: string]: any },
+    configuration: Configuration
+  ) => {
     const defaultContext = ContextGetDefaultService.call(
       context,
       '',